Add clang support for new Objective-C literal syntax for NSDictionary, NSArray, NSNumber, and boolean literals.  This includes both Sema and Codegen support. Included is also support for new Objective-C container subscripting.
My apologies for the large patch.  It was very difficult to break apart.
The patch introduces changes to the driver as well to cause clang to link
in additional runtime support when needed to support the new language features.

Docs are forthcoming to document the implementation and behavior of these features.