Add Vim frontend to LLDB. - Access any LLDB CLI command in Vim by typing ":L<command>". Tab-completion   works too! - See source locations for breakpoints and the current PC with vim "marks"   and highlights. - Examine backtraces, locals, disassembly, registers, and breakpoints in   dedicated Vim windows. - See when in-scope variables and registers change by watching for (red)   highlights.
This plugin opens multiple Vim "windows" to display debugger information.
To quit all windows at the same time use ":qa". The alternative would be
":q" to close each window separately.

This plugin is known to work on Mac OS X (Mountain Lion) with MacVim and
the system-provided terminal Vim, and on Linux (Ubuntu 12.04 and 12.10)
with GVim and the terminal Vim from the "vim-gnome" package.